// src/app/+core/interfaces/user.interface.ts
export interface IUser {
    email?: string;
    password?: string;
    passwordConfirmation?: string;
    fullName?: string;
    birthDate?: string;
    zipCode?: number;
    streetName?: string;
    number?: number;
    complement?: string;
    neighbourhood?: string;
    country?: string;
    state?: string;
    city?: string;
    token?: string;
}